Viktor

+++ account-data-server/src/main/java/via/sdj3/grpcserverexample/repository/OrderRepository.java
@@ -8,6 +8,6 @@ import via.sdj3.grpcserverexample.entities.UserEntity;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<OrderEntity, Long> {
+public interface OrderRepository extends JpaRepository<OrderEntity, Integer> {
 
 }
